specify usage of `httr` library on slide 65
Added a line stating that the `add_header()` function belongs to the library `httr`. This signals that they have to install and import the library, since the code will break on their scripts if `httr` is not imported (I simply copied the code on the slide and had to google multiple times until I found out why the code was breaking at `add_header`).

I did not provide the knitted html and pdf files to this commit because my knit is producing other line changes in other slides in the knitted document.